src/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
augly==1.0.0
autocommand==2.2.2
autofaiss==2.15.8
backports.csv==1.0.7
beautifulsoup4==4.11.2
black==23.3.0
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi
cffi
charset-normalizer
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colored==1.4.4
confection==0.0.4
contourpy==1.0.7
crcmod==1.7
cryptography
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
datasets==2.12.0
diffusers==0.14.0
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
embedding-reader==1.5.1
entrypoints==0.4
faiss-cpu==1.7.4
fastapi==0.95.1
feedparser==6.0.10
ffmpy==0.3.0
filelock==3.9.0
fire==0.4.0
flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.9
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gradio==3.27.0
gradio_client==0.1.3
grpcio==1.51.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
hydra-core==1.3.2
idna
imageio==2.27.0
importlib-metadata==6.8.0
importlib-resources==5.10.2
inflect==6.0.2
iopath==0.1.9
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
linkify-it-py==2.0.0
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.6.2
mmengine==0.8.4
mmsegmentation==0.27.0
model-index==0.1.11
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
mysqlclient==2.1.1
networkx==3.1
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.18.0
opencv-contrib-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.19
ordered-set==4.1.0
orjson==3.8.10
oss2==2.17.0
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
pathspec==0.11.1
pathtools==0.1.2
pathy==0.10.1
Pattern==3.6
pdfminer.six==20221105
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
portalocker==2.7.0
portend==3.1.0
preshed==3.0.8
prettytable==3.8.0
promise==2.3
protobuf==3.20.3
psutil==5.9.4
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser
pycryptodome==3.18.0
pydantic==1.10.4
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rsa==4.9
safetensors==0.3.2
scikit-image==0.20.0
scipy==1.9.1
semantic-version==2.10.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.2.2
sgmllib3k==1.0.0
shortuuid==1.0.8
six
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.26.1
tabulate==0.9.0
tempora==5.2.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thinc==8.1.9
tifffile==2023.4.12
timm==0.9.5
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.2
transformers==4.28.1
typer==0.7.0
typing_extensions
uc-micro-py==1.0.1
urllib3
uvicorn==0.21.1
visdom==0.1.8.9
wandb==0.15.8
wasabi==1.1.1
wcwidth==0.2.6
webdataset==0.2.33
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.2.2
xxhash==3.2.0
yacs==0.1.8
yapf==0.40.1
yarl==1.8.2
yaspin==2.1.0
zc.lockfile==2.0
zipp==3.13.0
